Unlock state and more tests, update deletes with refs failing
